% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{combine2Table}
\alias{combine2Table}
\title{Combine all IP and Input count table files}
\usage{
combine2Table(outputName, replicateNumber, currentDir)
}
\arguments{
\item{outputName}{Name of the output table}

\item{replicateNumber}{Number of the replicates}

\item{currentDir}{Directory for I/O operations}
}
\value{
No return value
}
\description{
Open raw counts IP and Inut files and with given total counts
 calculate Fold Enrichment values, and combine them into one file
}
