% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{combineMotifFiles}
\alias{combineMotifFiles}
\title{Combine motif bed files into a combined ranges}
\usage{
combineMotifFiles(motifFileNames, motifType = "BioString")
}
\arguments{
\item{motifFileNames}{a vector motif file names}

\item{motifType}{Type of motif string or give bed}
}
\value{
No return value
}
\description{
Get motif file names and combine them into a matrix, and keep
the indices of original motifs in the combined file.

If the motif type is string the bed files are deleted after
being combined to one matrix.
}
