% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{combineTestResults}
\alias{combineTestResults}
\title{Combine count Table and statistics table}
\usage{
combineTestResults(
  motifFile,
  acceptedMotifsOutputFile,
  acceptedRegionsOutputFile,
  countTableFile,
  testTableFile,
  fdrCutoff,
  windowSize
)
}
\arguments{
\item{motifFile}{File contains motifs}

\item{acceptedMotifsOutputFile}{File name of accepted motif table inforation}

\item{acceptedRegionsOutputFile}{File name of accepteted region information}

\item{countTableFile}{Table of count values file name}

\item{testTableFile}{negative binomial test table file name}

\item{fdrCutoff}{Pvalue cut-off related to the used FDR}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}
}
\value{
The average binding intensity for each ChIP-seq
}
\description{
Combine count table and pvalue FE statistics into one file for
 motifs and regions seperately.
}
