% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{CompareBeds2GivenRegions}
\alias{CompareBeds2GivenRegions}
\title{Compare a set of bed files to a provided regions set}
\usage{
CompareBeds2GivenRegions(motifName, bindingRegions)
}
\arguments{
\item{motifName}{a vector of motif names}

\item{bindingRegions}{granges of provided binding regions}
}
\value{
A dataframe which includes precision recall values for each motif
}
\description{
Get combined ranges of bed files and compare them to given

binding regions in terms of precision/recall.
}
