% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{CompareMotifs2GivenRegions}
\alias{CompareMotifs2GivenRegions}
\title{Comparison motifs locations to a given regions set}
\usage{
CompareMotifs2GivenRegions(motifs, mismatchNumbers, bindingRegions)
}
\arguments{
\item{motifs}{a vector of motif characters in nucleotide IUPAC format}

\item{mismatchNumbers}{a vector Number of mismatches allowed to match with
motifs}

\item{bindingRegions}{granges of user provided binding regions}
}
\value{
A dataframe which includes precision recall values for each motif
}
\description{
Comparison of motifs  locations to user provided binding
regions.

It returns this comparison to given user binding regions in
terms of precision/recall.
}
