% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{computeFoldEnrichment}
\alias{computeFoldEnrichment}
\title{compute fold enrichment values for an experiment}
\usage{
computeFoldEnrichment(
  ipCountFile,
  inputCountFile,
  ipTotalCount,
  inputTotalCount,
  outputName
)
}
\arguments{
\item{ipCountFile}{File contains motifs count values for IP experiment}

\item{inputCountFile}{File contains motifs count values for Input experiment}

\item{ipTotalCount}{Total short reads number in IP experiment}

\item{inputTotalCount}{Total short reads number in Input experiment}

\item{outputName}{Name of the output table}
}
\value{
No return value
}
\description{
Open raw counts IP and Inut files and with given total counts
 calculate Fold Enrichment values for the motifs
}
