% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{DeleteMultipleFiles}
\alias{DeleteMultipleFiles}
\title{Delete a vector of files}
\usage{
DeleteMultipleFiles(files)
}
\arguments{
\item{files}{a vector of files}
}
\value{
No return value
}
\description{
Delete multiple give files as a vector of characters
}
