% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{DetectBindingSitesBed}
\alias{DetectBindingSitesBed}
\title{Detect binding sites from bed motif input}
\usage{
DetectBindingSitesBed(
  BedFile,
  IPfiles,
  BackgroundFiles,
  genome,
  genomeBuild,
  fdrValue = 0.05,
  expName = "Motif_Centric_Peaks",
  windowSize = 100,
  format = ""
)
}
\arguments{
\item{BedFile}{Motif locations in bed format file}

\item{IPfiles}{IP ChIP-seq alignment files}

\item{BackgroundFiles}{Background ChIP-seq alignment files. Can be Input
experimetn, DNA whole exctract, etc.}

\item{genome}{The genome name such as "Hsapiens", "Mmusculus",
"Dmelanogaster"}

\item{genomeBuild}{The genome build such as "hg38", "hg19", "mm10", "dm3"}

\item{fdrValue}{FDR value cut-off}

\item{expName}{The name of the output table}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}

\item{format}{alignment format and should be one of these: "BAMPE", "BAMSE",
"BEDPE", "BEDSE"}
}
\value{
peakCallingStatistics A list FRiPs, sequence statistics, and Motif
 statistics
}
\description{
Takes user provied bed regions, and check for validity of them.
 Read bam or bed alignment files and convert to 1 nt bed and call detect
  binding site from 1nt bed.
}
\examples{

# HOX candidate motifs in Chr19 mouse
HOXMotifs = system.file("extdata", "HOXHomer_chr19.bed",
                         package="Motif2Site")

# HOXA2 BA2 peak calling from bed file
# ChIP-seq datasets in bam paired end format
IPs <- c(system.file("extdata", "HOXA2_BA2_rep1_IP_chr19.bam",
                     package="Motif2Site"),
        system.file("extdata", "HOXA2_BA2_rep2_IP_chr19.bam",
                    package="Motif2Site"))
INPUTs <- c(system.file("extdata", "HOXA2_BA2_rep1_INPUT_chr19.bam",
                        package="Motif2Site"),
            system.file("extdata", "HOXA2_BA2_rep2_INPUT_chr19.bam",
                        package="Motif2Site"))
HOXA2BA2_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BAMPE" ,
                                             expName = "HOXA2_BA2"
                                             )

# HOXA3 PBA2 peak calling from bed file motif
# ChIP-seq datasets in bed single end format
IPs <- c(system.file("extdata", "HOXA3_PBA_rep1_IP_chr19.bed.bz2",
                     package="Motif2Site"),
         system.file("extdata", "HOXA3_PBA_rep2_IP_chr19.bed.bz2",
                     package="Motif2Site"))
INPUTs <- c(system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"),
            system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"))
HOXA3PBA_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BEDSE",
                                             expName = "HOXA3_PBA"
                                             )

# HOXA2 PBA peak calling from bed file motif
# ChIP-seq datasets in bam paired end format
IPs <- system.file("extdata","HOXA2_PBA_IP_chr19.bam",
                   package="Motif2Site")
INPUTs <- system.file("extdata","HOXA2_PBA_INPUT_chr19.bam",
                      package="Motif2Site")
HOXA2PBA_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BAMPE",
                                             expName = "HOXA2_PBA",
                                             fdrValue = 0.1
                                             )

}
\seealso{
\code{\link{DetectBindingSitesMotif}}
}
