% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{DetectFdrCutoffBH}
\alias{DetectFdrCutoffBH}
\title{FDR cut-off detection Benjamini Hochberg method}
\usage{
DetectFdrCutoffBH(TestTableFile = "TestResults", fdrValue = 0.05)
}
\arguments{
\item{TestTableFile}{test table which contains pvalues}

\item{fdrValue}{FDR cut-off}
}
\value{
pvalue cut-off
}
\description{
Return FDR cut-off for a user provided fdrvalue using Benjamini
 Hochberg on main motif test data
}
