% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{fitKernelDensity}
\alias{fitKernelDensity}
\title{Fit a kernel density distribution to the obersever heuristic
 distribution}
\usage{
fitKernelDensity(heuristicDistribution, totalShortReads, windowSize)
}
\arguments{
\item{heuristicDistribution}{Original short distribution}

\item{totalShortReads}{Total number of short reads}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}
}
\value{
kernel returns fitted kernel distribution of short reads around
 binding sites
}
\description{
This function gets heuristic distribution of short reads around
 binding sites, total numer of short reads and window size as number of
  neucleotid around binding sites.

It fits a kernel to the distribution and return the distribution
 as output. The total sum of returned values is equal to one.
}
