% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{generate1ntBedAlignment}
\alias{generate1ntBedAlignment}
\title{Convert bam and bed files to 1 nucleotide bed}
\usage{
generate1ntBedAlignment(InputFile, bedFile, format = "")
}
\arguments{
\item{InputFile}{Original alignment file name}

\item{bedFile}{Name of output 1nt bed file}

\item{format}{alignment format and should be one of these: "BAMPE", "BAMSE",
"BEDPE", "BEDSE"}
}
\value{
No return value
}
\description{
Take alignment files in bam or bed fomat and convert them to 1
 nucleotide bed file
}
