% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{Granges2Bed}
\alias{Granges2Bed}
\title{Write Genomic Ranges in a bed file}
\usage{
Granges2Bed(granges, fileName)
}
\arguments{
\item{granges}{coordinates to write in granges format}

\item{fileName}{the name of the file that contains the coordianted}
}
\value{
No return value
}
\description{
Write Genomic Ranges in a bed file.
}
