% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{motifBindingNegativeBinomialCount}
\alias{motifBindingNegativeBinomialCount}
\title{Model IP and Input count values with negative Binomal}
\usage{
motifBindingNegativeBinomialCount(
  countTableFile,
  replicateNumber,
  outputFile,
  currentDir
)
}
\arguments{
\item{countTableFile}{Table of counts which contains all IP and Input value
raw counts}

\item{replicateNumber}{experiment replicate number}

\item{outputFile}{The name of the output file generated by this function}

\item{currentDir}{Directory for I/O operations}
}
\value{
A dataframe includes fold enrichment, pvalue, and normalized count
 values
}
\description{
Using edgeR TMM normalization and estimating dispersion as well
 as Adapting exact test function from edgeR to model IP vs Input counts.

To make this function memory effcient motifs into smaller sets
 and compute them seperately and combine them at the end.
}
