% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{motifChipCount}
\alias{motifChipCount}
\title{count short reads related to each motif for a given ChIPseq file}
\usage{
motifChipCount(motifFile, chipFile, windowSize, outputName)
}
\arguments{
\item{motifFile}{File contains motifs}

\item{chipFile}{ChIP-seq 1nt alignment locations in bed format}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}

\item{outputName}{Name of the output table}
}
\value{
Total number of short reads in motif reagions
}
\description{
count 1nt short reads related to each motif for a given ChIPseq
 file.
}
