% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{motifCount}
\alias{motifCount}
\title{count short reads around motifs for all ChIP-seq experiments}
\usage{
motifCount(motifFile, chipSeq, windowSize, outputName, currentDir)
}
\arguments{
\item{motifFile}{File contains motifs}

\item{chipSeq}{dataframe of ChIP-seq 1nt alignment location}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}

\item{outputName}{Name of the output table}

\item{currentDir}{Directory for I/O operations}
}
\value{
No return value
}
\description{
count short reads related to each motif for all ChIPseq files
 both IP and Input.
}
