% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{motifTablePreProcess}
\alias{motifTablePreProcess}
\title{Process count data and perform negative binomial test}
\usage{
motifTablePreProcess(countTableFile, outFile, currentDir)
}
\arguments{
\item{countTableFile}{Tabl of count values around motifs for all ChIP-seq
experiments}

\item{outFile}{The name of the output file}

\item{currentDir}{Directory for I/O operations}
}
\value{
sequencingStatitics A dataframe consists of the ratio of
 non-sequenced, low-sequenced, ang high-sequenced regions.
}
\description{
Remove unmmaped regions, low and high binding regions and
 regions without fold change, and call negative binomial or nb test for
  the remaining regions.
}
