% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{NegativeBinomialTestWithReplicate}
\alias{NegativeBinomialTestWithReplicate}
\title{Negative binomial test of binding using all replicates}
\usage{
NegativeBinomialTestWithReplicate(object, prior.count = 0.125)
}
\arguments{
\item{object}{Table of counts which contains all IP and Input value counts,
TMM normalized and contains dispersion values}

\item{prior.count}{edgeR prior value}
}
\value{
log fold enrichment, pvalue, and normalized count values
}
\description{
Adapted exact test function from edgeR to compare IP vs Input
 with replicates. Input is a DGELIST with common and tag-wise dispression has
  been already caluclated by edgeR commands.

It calculates abundaces with mglmOneGroup identical to edgeR.
 logFE was  calculated identiacl to edgeR. For the pvalue test negative
  binomial test is performed on the calculated abundance.
}
