% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{pairwisDifferential}
\alias{pairwisDifferential}
\title{Detect differential motifs}
\usage{
pairwisDifferential(
  tableOfCountsDir = "",
  exp1,
  exp2,
  FDRcutoff = 0.05,
  logFCcuttoff = 1
)
}
\arguments{
\item{tableOfCountsDir}{Directory which conatins the combined motifs and
ChIP-seq count file}

\item{exp1}{Experiment name which will be compared in pairwise comparison}

\item{exp2}{Experiment name which will be compared in pairwise comparison}

\item{FDRcutoff}{FDR cutoff applies on pvalue distribution}

\item{logFCcuttoff}{log fold change cutoff}
}
\value{
A list of differential motifs, motif1 and motif2 as well as a table
 of total motifs and log fold changes
}
\description{
Take combined matrix of motif counts generated by
 \code{\link{recenterBindingSitesAcrossExperiments}}, and experiment names.
 It detect differential motifs using edgeR TMM nomralizaiton with Generalized
  linear model
}
\examples{

# HOX candidate motifs in Chr19 mouse
HOXMotifs = system.file("extdata", "HOXHomer_chr19.bed",
                        package="Motif2Site")

# HOXA2 BA2 peak calling from bed file
# ChIP-seq datasets in bam paired end format
IPs <- c(system.file("extdata", "HOXA2_BA2_rep1_IP_chr19.bam",
                     package="Motif2Site"),
         system.file("extdata", "HOXA2_BA2_rep2_IP_chr19.bam",
                     package="Motif2Site"))
INPUTs <- c(system.file("extdata", "HOXA2_BA2_rep1_INPUT_chr19.bam",
                        package="Motif2Site"),
           system.file("extdata", "HOXA2_BA2_rep2_INPUT_chr19.bam",
                       package="Motif2Site"))
HOXA2BA2_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BAMPE",
                                             expName = "HOXA2_BA2"
                                             )

# HOXA3 PBA2 peak calling from bed file motif
# ChIP-seq datasets in bed single end format
IPs <- c(system.file("extdata", "HOXA3_PBA_rep1_IP_chr19.bed.bz2",
                     package="Motif2Site"),
         system.file("extdata", "HOXA3_PBA_rep2_IP_chr19.bed.bz2",
                     package="Motif2Site"))
INPUTs <- c(system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"),
            system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"))
HOXA3PBA_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BEDSE",
                                             expName = "HOXA3_PBA"
                                             )

# HOXA2 PBA peak calling from bed file motif
# ChIP-seq datasets in bam paired end format
IPs <- system.file("extdata","HOXA2_PBA_IP_chr19.bam",
                   package="Motif2Site")
INPUTs <- system.file("extdata","HOXA2_PBA_INPUT_chr19.bam",
                      package="Motif2Site")
HOXA2PBA_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BAMPE",
                                             expName = "HOXA2_PBA",
                                              fdrValue = 0.1
                                              )

# Combine all HOX binding sites into one table
corMAT <- recenterBindingSitesAcrossExperiments(
  expLocations=c("HOXA2_BA2","HOXA2_PBA", "HOXA3_PBA"),
  experimentNames=c("HOXA2BA2","HOXA2PBA", "HOXA3PBA"),
  expName="combinedHOX",
  fdrValue=0.05
  )

# Differential binding sites across HOXA2_BA2 and HOXA3_PBA
diffHOX <- pairwisDifferential(tableOfCountsDir="combinedHOX",
                               exp1="HOXA2BA2",
                               exp2="HOXA3PBA",
                               FDRcutoff = 0.05,
                               logFCcuttoff = 1
                               )
HOXA2BA2up <- diffHOX[[1]]
HOXA3PBAup <- diffHOX[[2]]
TotalComparison <- diffHOX[[3]]
head(TotalComparison)

}
\seealso{
\code{\link{recenterBindingSitesAcrossExperiments}}
}
