% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{quiet}
\alias{quiet}
\title{Suppress messages generated by in external package}
\usage{
quiet(func)
}
\arguments{
\item{func}{functional input call for which cat messages should be supressed}
}
\value{
No return value
}
\description{
mixtools and MASS::fitdistr generates warning by cat which is
suppressed by this funcitons
}
