% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{recenterBindingSitesAcrossExperiments}
\alias{recenterBindingSitesAcrossExperiments}
\title{Combine binding sites across experiments}
\usage{
recenterBindingSitesAcrossExperiments(
  expLocations,
  experimentNames,
  expName = "combinedData",
  fdrValue = 0.05,
  fdrCrossExp = 0.001
)
}
\arguments{
\item{expLocations}{The path to the experiment folders}

\item{experimentNames}{Name of the experiment to be used in combined ChIP-seq}

\item{expName}{Name of the combined matrix}

\item{fdrValue}{FDR cut-off to accept binding in each ChIP-seq experiments}

\item{fdrCrossExp}{If no experiment fullfill this cutoff, the motif is not
considered}
}
\value{
A pariwise Pearson correlation matrix across experiments
}
\description{
Take experiment folder locations and experiment names and
 combine them into a combined matrix of motifs and ChIP-seq counts

Experiment folders must be generated either by
\code{\link{DetectBindingSitesBed}} or \code{\link{DetectBindingSitesMotif}}.
}
\examples{

# HOX candidate motifs in Chr19 mouse
HOXMotifs = system.file("extdata", "HOXHomer_chr19.bed",
                        package="Motif2Site")

# HOXA2 BA2 peak calling from bed file
# ChIP-seq datasets in bam paired end format
IPs <- c(system.file("extdata", "HOXA2_BA2_rep1_IP_chr19.bam",
                     package="Motif2Site"),
         system.file("extdata", "HOXA2_BA2_rep2_IP_chr19.bam",
            package="Motif2Site")
         )
INPUTs <- c(system.file("extdata", "HOXA2_BA2_rep1_INPUT_chr19.bam",
                        package="Motif2Site"),
            system.file("extdata", "HOXA2_BA2_rep2_INPUT_chr19.bam",
            package="Motif2Site")
            )
HOXA2BA2_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BAMPE",
                                             expName = "HOXA2_BA2"
                                             )

# HOXA3 PBA2 peak calling from bed file motif
# ChIP-seq datasets in bed single end format
IPs <- c(system.file("extdata", "HOXA3_PBA_rep1_IP_chr19.bed.bz2",
                     package="Motif2Site"),
         system.file("extdata", "HOXA3_PBA_rep2_IP_chr19.bed.bz2",
                     package="Motif2Site")
         )
INPUTs <- c(system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"),
            system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site")
            )
HOXA3PBA_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BEDSE",
                                             expName = "HOXA3_PBA"
                                             )

# HOXA2 PBA peak calling from bed file motif
# ChIP-seq datasets in bam paired end format
IPs <- system.file("extdata","HOXA2_PBA_IP_chr19.bam", package="Motif2Site")
INPUTs <- system.file("extdata","HOXA2_PBA_INPUT_chr19.bam",
                      package="Motif2Site")
HOXA2PBA_statistics <- DetectBindingSitesBed(BedFile= HOXMotifs,
                                             IPfiles = IPs,
                                             BackgroundFiles = INPUTs,
                                             genome="Mmusculus",
                                             genomeBuild="mm10",
                                             format="BAMPE",
                                             expName = "HOXA2_PBA",
                                              fdrValue = 0.1
                                              )

# Combine all HOX binding sites into one table
corMAT <- recenterBindingSitesAcrossExperiments(
  expLocations=c("HOXA2_BA2","HOXA2_PBA", "HOXA3_PBA"),
  experimentNames=c("HOXA2BA2","HOXA2PBA", "HOXA3PBA"),
  expName="combinedHOX",
  fdrValue=0.05
  )
corMAT
matFile <- paste0(getwd(), "/combinedHOX/CombinedMatrix")
HOXTable <- read.table(matFile, header = TRUE, check.names = FALSE)
head(HOXTable)

}
\seealso{
\code{\link{pairwisDifferential}}
}
