% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{strongestMotif}
\alias{strongestMotif}
\title{Returns the motif with the highest count}
\usage{
strongestMotif(motifLocations, readLocations, windowSize)
}
\arguments{
\item{motifLocations}{A vector of motif locations}

\item{readLocations}{A vector of 1nt short reads}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}
}
\value{
The strongest motif
}
\description{
Gets motif locations and related short reads and returns the
 motif which include the highest number of short reads around it.
}
