% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif2Site.R
\name{DetectBindingSitesMotif}
\alias{DetectBindingSitesMotif}
\title{Detect binding sites from sequence motif sequence and mismatchNumber}
\usage{
DetectBindingSitesMotif(
  motif,
  mismatchNumber,
  IPfiles,
  BackgroundFiles,
  genome,
  genomeBuild,
  fdrValue = 0.05,
  expName = "Motif_Centric_Peaks",
  windowSize = 100,
  format = "",
  GivenRegion = NA
)
}
\arguments{
\item{motif}{motif characters in nucleotide IUPAC format}

\item{mismatchNumber}{Number of mismatches allowed to match with motifs}

\item{IPfiles}{IP ChIP-seq alignment files}

\item{BackgroundFiles}{Background ChIP-seq alignment files. Can be Input
experimetn, DNA whole exctract, etc.}

\item{genome}{The genome name such as "Hsapiens", "Mmusculus",
"Dmelanogaster"}

\item{genomeBuild}{The genome build such as "hg38", "hg19", "mm10", "dm3"}

\item{fdrValue}{FDR value cut-off}

\item{expName}{The name of the output table}

\item{windowSize}{Window size around binding site. The total region would be
2*windowSize+1}

\item{format}{alignment format and should be one of these: "BAMPE", "BAMSE",
"BEDPE", "BEDSE"}

\item{GivenRegion}{granges of user provided binding regions}
}
\value{
A list FRiPs, sequence statistics, and Motif statistics
}
\description{
DETECT Binding sites with given motif and mismatch number as
well genome/build, False Discovery Rate for a given experiment name. Read bam
 or bed alignment files and convert to 1 nt bed and detect binding site among
  motifs from 1nt bed alignment.
}
\examples{

# HOXA2 BA2 peak calling from string motif
# ChIP-seq datasets in bam paired end format
IPs <- c(system.file("extdata", "HOXA2_BA2_rep1_IP_chr19.bam",
                     package="Motif2Site"),
         system.file("extdata", "HOXA2_BA2_rep2_IP_chr19.bam",
                     package="Motif2Site"))
INPUTs <- c(system.file("extdata", "HOXA2_BA2_rep1_INPUT_chr19.bam",
                        package="Motif2Site"),
            system.file("extdata", "HOXA2_BA2_rep2_INPUT_chr19.bam",
                        package="Motif2Site"))
HOXA2BA2_statistics <- DetectBindingSitesMotif(motif= "TGATNNAT",
                                               mismatchNumber=1,
                                               IPfiles = IPs,
                                               BackgroundFiles = INPUTs,
                                               genome="Mmusculus",
                                               genomeBuild="mm10",
                                               format="BAMPE",
                                               expName = "HOXA2_BA2_String"
                                               )

# HOXA3 PBA2 peak calling from bed file motif
# ChIP-seq datasets in bed single end format
# peak calling only on motifs which are interesected by MEIS binding reigon
IPs <- c(system.file("extdata", "HOXA3_PBA_rep1_IP_chr19.bed.bz2",
                     package="Motif2Site"),
         system.file("extdata", "HOXA3_PBA_rep2_IP_chr19.bed.bz2",
                     package="Motif2Site"))
INPUTs <- c(system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"),
            system.file("extdata", "HOXA3_PBA_rep1_INPUT_chr19.bed.bz2",
                        package="Motif2Site"))
Meis_chr19 <- Bed2Granges(system.file("extdata", "Meis_Chr19.bed",
                                      package="Motif2Site"))
HOXA3PBA_statistics <- DetectBindingSitesMotif(motif= "TGATNNAT",
                                               mismatchNumber=1,
                                               IPfiles = IPs,
                                               BackgroundFiles = INPUTs,
                                               genome="Mmusculus",
                                               genomeBuild="mm10",
                                               format="BEDSE",
                                               expName = "HOXA3_PBA_String",
                                               GivenRegion = Meis_chr19
                                               )

}
\seealso{
\code{\link{DetectBindingSitesBed}}
}
